[Refactor] 테라리움 응답 수정 + HTTP 에러 상태 코드 수정 + 쓰지않는 라이브러리 import 삭제 + 폴더 구조 통일
