+++ src/main/java/umc/plantory/domain/diary/dto/DiaryRequestDTO.java
@@ -6,8 +6,6 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
 import umc.plantory.global.validation.annotation.ValidDiaryFields;
 
 import java.time.LocalDate;
