+++ src/main/java/umc/plantory/domain/flower/repository/FlowerRepository.java
@@ -4,8 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import umc.plantory.domain.flower.entity.Flower;
 import umc.plantory.global.enums.Emotion;
 
-import java.util.Optional;
-
 public interface FlowerRepository extends JpaRepository<Flower, Long> {
     Flower findByEmotion(Emotion emotion);
 }
