+++ src/main/java/umc/plantory/domain/member/dto/MemberDataDTO.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import umc.plantory.global.enums.Gender;
 
 public class MemberDataDTO {
