+++ src/main/java/umc/plantory/domain/member/service/MemberQueryService.java
@@ -3,11 +3,8 @@ package umc.plantory.domain.member.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import umc.plantory.domain.diary.dto.DiaryResponseDTO;
 import umc.plantory.domain.diary.entity.Diary;
 import umc.plantory.domain.diary.repository.DiaryRepository;
-import umc.plantory.domain.diary.service.DiaryQueryUseCase;
-import umc.plantory.domain.flower.entity.Flower;
 import umc.plantory.domain.member.converter.MemberConverter;
 import umc.plantory.domain.member.dto.MemberResponseDTO;
 import umc.plantory.domain.member.entity.Member;
