+++ src/main/java/umc/plantory/domain/terrarium/converter/TerrariumConverter.java
@@ -68,10 +68,10 @@ public class TerrariumConverter {
                 .build();
     }
 
-    public static TerrariumResponseDto.CompletedTerrariumDetatilResponse toCompletedTerrariumDetatilResponse(Terrarium terrarium, List<LocalDate> usedDiaries) {
-        return TerrariumResponseDto.CompletedTerrariumDetatilResponse.builder()
-                .startAt(terrarium.getStartAt())
-                .bloomAt(terrarium.getBloomAt())
+    public static TerrariumResponseDto.CompletedTerrariumDetailResponse toCompletedTerrariumDetatilResponse(Terrarium terrarium, List<LocalDate> usedDiaries) {
+        return TerrariumResponseDto.CompletedTerrariumDetailResponse.builder()
+                .startAt(terrarium.getStartAt().toLocalDate())
+                .bloomAt(terrarium.getBloomAt().toLocalDate())
                 .mostEmotion(terrarium.getFlower().getEmotion())
                 .firstStepDate(terrarium.getFirstStepDate())
                 .secondStepDate(terrarium.getSecondStepDate())
