+++ src/main/java/umc/plantory/domain/terrarium/entity/Terrarium.java
@@ -4,10 +4,8 @@ import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.DynamicUpdate;
-import org.hibernate.annotations.Fetch;
 import umc.plantory.domain.flower.entity.Flower;
 import umc.plantory.domain.member.entity.Member;
-import umc.plantory.global.baseEntity.BaseEntity;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
