+++ src/main/java/umc/plantory/domain/terrarium/repository/TerrariumRepository.java
@@ -5,8 +5,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import umc.plantory.domain.terrarium.entity.Terrarium;
 import umc.plantory.domain.member.entity.Member;
-import umc.plantory.domain.terrarium.entity.Terrarium;
-
 import java.util.List;
 import java.util.Optional;
